+++ src/app/routes/accounts.ts
@@ -222,7 +222,21 @@ accountsRouter.get(
                 page: (req.query.page !== undefined) ? Math.max(req.query.page, 1) : 1,
                 accountNumber: req.params.accountNumber
             };
-            const actions = await actionRepo.searchTransferActions(searchConditions);
+            let actions = await actionRepo.searchTransferActions(searchConditions);
+
+            // 互換性維持対応
+            actions = actions.map((a) => {
+                return {
+                    ...a,
+                    amount: (typeof a.amount === 'number')
+                        ? {
+                            typeOf: 'MonetaryAmount',
+                            currency: 'Point', // 旧データはPointしかないのでこれで十分
+                            value: a.amount
+                        }
+                        : a.amount
+                };
+            });
 
             res.json(actions);
         } catch (error) {